[refactor] 랭크요약컴포넌트를 left-side 디렉토리로 변경
