+++ src/components/left-side/TierBox.svelte
