+++ src/components/left-side/index.ts
@@ -1,2 +1,3 @@
 export { default as TierBox } from './TierBox.svelte';
 export { default as SubTierBox } from './SubTierBox.svelte';
+export { default as LastSevenDays } from './LastSevenDays/LastSevenDays.svelte';
